requirements.txt
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.41.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
pandas>=2.2.0
